// types/nutritionTypes.ts
export interface NutritionProps {
  id: number;
  name: string;
  cal: string;
  pro: string;
  fat: string;
  carb: string;
  fibr: string;
  type: string;
}
export interface CookProps {
  id: number;
  title: string;
  media: string;
  content1: string;
  content2: string;
}

export interface CategoryProps {
  id: number;
  title: string;
  img: string;
  slug?: string;
}

export interface NutritionProps {
  id: number;
  name: string;
  cal: string;
  pro: string;
  fat: string;
  carb: string;
  fibr: string;
  type: string;
}